Rename `Types.js` to `index.js` under FlowTypes & edit import statements
