+++ src/Components/LoginView.jsx
@@ -9,7 +9,7 @@ import logo from '../images/slack-logo.png';
 // import slackConnectHref from '../Constants/connectWithSlackHref';
 import slack from '../images/slackIcon.png';
 import { connectWithSlack } from '../Actions/index';
-import type { Dispatch } from '../FlowTypes/Types';
+import type { Dispatch } from '../FlowTypes/';
 
 export function LoginView(props: Object) {
   const { connectWithSlack } = props;
