+++ src/Components/Message.jsx
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import { Image, List } from 'semantic-ui-react';
-import type { MessageType } from '../FlowTypes/Types';
+import type { MessageType } from '../FlowTypes/';
 
 
 export default function Message(
