+++ src/index.js
@@ -18,8 +18,8 @@ ReactDOM.render(
 );
 registerServiceWorker();
 
-// const socket = openSocket('http://localhost:4000');
-const socket = openSocket('https://databraid.localtunnel.me');
+ const socket = openSocket('http://localhost:4000');
+//const socket = openSocket('https://databraid.localtunnel.me');
 
 
 socket.on('messages', (messages) => {
